Backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
celery==5.3.0
redis==4.5.0
Pillow==9.5.0
requests==2.31.0
djangorestframework-simplejwt==5.3.1
